2/23/26 Fix

+++ xaga/sources/core/arm64/atomic_explicit.cxx
@@ -182,7 +182,7 @@ namespace rainy::core::pal {
     bool interlocked_compare_exchange_explicit(volatile long *destination, long exchange, long comparand, memory_order success,
                                                memory_order failure) {
         bool result{};
-        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange, result, success, reinterpret_cast<volatile std::int64_t *>(destination),
+        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange, result, success, reinterpret_cast<volatile std::int32_t *>(destination),
                               exchange, comparand);
         return result;
     }
@@ -190,7 +190,7 @@ namespace rainy::core::pal {
     bool interlocked_compare_exchange8_explicit(volatile std::int8_t *destination, std::int8_t exchange, std::int8_t comparand,
                                                 memory_order success, memory_order failure) {
         bool result{};
-        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange8, result, success, reinterpret_cast<volatile std::int64_t *>(destination),
+        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange8, result, success, reinterpret_cast<volatile std::int8_t *>(destination),
                               exchange, comparand);
         return result;
     }
@@ -198,7 +198,7 @@ namespace rainy::core::pal {
     bool interlocked_compare_exchange16_explicit(volatile std::int16_t *destination, std::int16_t exchange, std::int16_t comparand,
                                                  memory_order success, memory_order failure) {
         bool result{};
-        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange16, result, success, reinterpret_cast<volatile std::int64_t *>(destination),
+        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange16, result, success, reinterpret_cast<volatile std::int16_t *>(destination),
                               exchange, comparand);
         return result;
     }
@@ -206,7 +206,7 @@ namespace rainy::core::pal {
     bool interlocked_compare_exchange32_explicit(volatile std::int32_t *destination, std::int32_t exchange, std::int32_t comparand,
                                                  memory_order success, memory_order failure) {
         bool result{};
-        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange32, result, success, reinterpret_cast<volatile std::int64_t *>(destination),
+        RAINY_ATOMIC_DISPATCH(_InterlockedCompareExchange32, result, success, reinterpret_cast<volatile std::int32_t *>(destination),
                               exchange, comparand);
         return result;
     }
